Put tasklist table into separate component
So we can render tasklists as tables without the filters and progress tabs, but without having to redefine all the formatters.
